01.03.2026 (2)

+++ 1_praktinis_darbas/src/main/java/lt/viko/eif/dalencinovic/first/spring/db/CarDealershipRepository.java
@@ -3,5 +3,8 @@ package lt.viko.eif.dalencinovic.first.spring.db;
 import lt.viko.eif.dalencinovic.first.spring.model.CarDealership;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarDealershipRepository extends JpaRepository<CarDealership,Long> {
+/**
+ * Repository interface for CarDealership entity.
+ */
+public interface CarDealershipRepository extends JpaRepository<CarDealership,Integer> {
 }
